backend/requirements.txt
fastapi>=0.116,<1
uvicorn[standard]>=0.35,<1
pydantic>=2.12
pydantic-settings>=2.0
httpx>=0.27
openai>=1.101
python-multipart>=0.0.9
reportlab>=4.2
pypdf>=4.3
python-pptx>=1.0
Pillow>=10.4
pytest>=8.3
pytest-cov>=5.0
pytest-asyncio>=0.23
